#19: Improved sdk selection logic in run config to auto-pick the project default ahk sdk if it exists, or otherwise pick any existing Ahk sdk in the sdk list.
