+++ src/main/resources/localization/AhkBundle.properties
@@ -9,7 +9,7 @@ runconfig.configtab.scriptargs.label=Script &Arguments:
 runconfig.configtab.scriptrunner.label=Script &Runner:
 runconfig.configtab.scriptrunner.sdklabel.default=Project Default
 runconfig.configtab.scriptrunner.sdklabel.unrecognized=Unrecognized SDK
-runconfig.configtab.scriptrunner.sdklabel.noneselected=No AutoHotkey SDK selected
+runconfig.configtab.scriptrunner.sdklabel.noneselected=No AutoHotkey SDK found
 
 runconfig.configtab.error.scriptpath.notexist=The script path does not exist on disk!
 runconfig.configtab.error.scriptpath.notahkextension=The script path does not point to an AutoHotkey (.ahk) file!
